Updates to Scripts 0, 8 and Module 8
Script 8 and Module 8:  Continue to build out the KNN algorithm.  We now have a working KNN algorithm that will return results for a range of nearest neighbor nodes.

In addition, I am working on a function that will generate the prediction score for different years.  This will help to show if there is a relationship between the year that a lawsuit was filed and the predictability of our algorithm.